# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/fibonacci/iterative/{n}": {
            "get": {
                "description": "Get the fibonacci number with iterative algorithm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fibonacci"
                ],
                "summary": "Get the fibonacci number with iterative algorithm",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number",
                        "name": "n",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/schemas.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/schemas.BaseResponse"
                        }
                    }
                }
            }
        },
        "/fibonacci/recursive/{n}": {
            "get": {
                "description": "Get the fibonacci number with recursive algorithm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fibonacci"
                ],
                "summary": "Get the fibonacci number with recursive algorithm",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number",
                        "name": "n",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/schemas.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/schemas.BaseResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Responds if the server is alive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Responds if the server is alive",
                "responses": {
                    "200": {
                        "description": "BaseResponse",
                        "schema": {
                            "$ref": "#/definitions/schemas.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schemas.BaseResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}